# packages/api-contracts/openapi/workplace.openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/groups": {
      "get": {
        "operationId": "GroupsController_listGroups",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Groups"
        ]
      },
      "post": {
        "operationId": "GroupsController_createGroup",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateGroupDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "Groups"
        ]
      }
    },
...
    "/groups/{groupId}/files": {
      "get": {
        "operationId": "GroupCollaborationController_listFiles",
        "parameters": [
          {
            "name": "groupId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "GroupCollaboration"
        ]
      },
      "post": {
        "operationId": "GroupCollaborationController_createFile",
        "parameters": [
          {
            "name": "groupId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateGroupFileDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "GroupCollaboration"
        ]
      }
    },
    "/groups/{groupId}/files/{fileId}/download": {
      "get": {
        "operationId": "GroupCollaborationController_getFileDownload",
        "parameters": [
          {
            "name": "groupId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "fileId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "GroupCollaboration"
        ]
      }
    },
    "/groups/{groupId}/files/{fileId}": {
      "delete": {
        "operationId": "GroupCollaborationController_deleteFile",
        "parameters": [
          {
            "name": "groupId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "fileId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "GroupCollaboration"
        ]
      }
    },
    "/groups/{groupId}/albums": {
      "get": {
        "operationId": "GroupCollaborationController_listAlbums",
        "parameters": [
          {
            "name": "groupId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "GroupCollaboration"
        ]
      },
      "post": {
        "operationId": "GroupCollaborationController_createAlbum",
        "parameters": [
          {
            "name": "groupId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateGroupAlbumDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "GroupCollaboration"
        ]
      }
    },
    "/groups/{groupId}/albums/{albumId}": {
      "get": {
        "operationId": "GroupCollaborationController_getAlbum",
        "parameters": [
          {
            "name": "groupId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "albumId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "GroupCollaboration"
        ]
      }
    },
    "/groups/{groupId}/albums/{albumId}/photos": {
      "post": {
        "operationId": "GroupCollaborationController_addAlbumPhoto",
        "parameters": [
          {
            "name": "groupId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "albumId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AddGroupAlbumPhotoDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "GroupCollaboration"
        ]
      }
    },
    "/groups/{groupId}/albums/{albumId}/photos/{photoId}/download": {
      "get": {
        "operationId": "GroupCollaborationController_getAlbumPhotoDownload",
        "parameters": [
          {
            "name": "groupId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "albumId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "photoId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "GroupCollaboration"
        ]
      }
    },
    "/groups/{groupId}/albums/{albumId}/photos/{photoId}": {
      "delete": {
        "operationId": "GroupCollaborationController_deleteAlbumPhoto",
        "parameters": [
          {
            "name": "groupId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "albumId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "photoId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "GroupCollaboration"
        ]
      }
    },
    "/groups/{groupId}/events": {
      "get": {
        "operationId": "GroupCollaborationController_listEvents",
        "parameters": [
          {
            "name": "groupId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "GroupCollaboration"
        ]
      },
      "post": {
        "operationId": "GroupCollaborationController_createEvent",
        "parameters": [
          {
            "name": "groupId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateGroupEventDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "GroupCollaboration"
        ]
      }
    },
    "/groups/{groupId}/events/{eventId}": {
      "patch": {
        "operationId": "GroupCollaborationController_updateEvent",
        "parameters": [
          {
            "name": "groupId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "eventId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateGroupEventDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "GroupCollaboration"
        ]
      },
      "delete": {
        "operationId": "GroupCollaborationController_deleteEvent",
        "parameters": [
          {
            "name": "groupId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "eventId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "GroupCollaboration"
        ]
      }
    },
    "/groups/{groupId}/events/{eventId}/rsvp": {
      "put": {
        "operationId": "GroupCollaborationController_setEventRsvp",
        "parameters": [
          {
            "name": "groupId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "eventId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpsertGroupEventRsvpDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "GroupCollaboration"
        ]
      }
    },
    "/groups/{groupId}/events/{eventId}/calendar": {
      "get": {
        "operationId": "GroupCollaborationController_exportEventCalendar",
        "parameters": [
          {
            "name": "groupId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "eventId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "GroupCollaboration"
        ]
      }
    },
...
  },
  "info": {
    "title": "Workplace API",
    "description": "Workplace backend API documentation",
    "version": "0.1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "access-token": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
...
      "CreateGroupDto": {
        "type": "object",
        "properties": {}
      },
...
      "CreateGroupFileDto": {
        "type": "object",
        "properties": {
          "mediaId": {
            "type": "string",
            "format": "uuid",
            "description": "A scanned MediaObject scoped to this group."
          },
          "title": {
            "type": "string",
            "maxLength": 160
          }
        },
        "required": [
          "mediaId"
        ]
      },
      "CreateGroupAlbumDto": {
        "type": "object",
        "properties": {
          "title": {
            "type": "string",
            "minLength": 1,
            "maxLength": 120
          },
          "description": {
            "type": "string",
            "maxLength": 500
          }
        },
        "required": [
          "title"
        ]
      },
      "AddGroupAlbumPhotoDto": {
        "type": "object",
        "properties": {
          "mediaId": {
            "type": "string",
            "format": "uuid",
            "description": "A scanned image MediaObject scoped to this group."
          },
          "caption": {
            "type": "string",
            "maxLength": 500
          }
        },
        "required": [
          "mediaId"
        ]
      },
      "CreateGroupEventDto": {
        "type": "object",
        "properties": {
          "title": {
            "type": "string",
            "minLength": 1,
            "maxLength": 160
          },
          "description": {
            "type": "string",
            "maxLength": 2000
          },
          "location": {
            "type": "string",
            "maxLength": 200
          },
          "startsAt": {
            "type": "string",
            "format": "date-time"
          },
          "endsAt": {
            "type": "string",
            "format": "date-time"
          },
          "timezone": {
            "type": "string",
            "example": "Asia/Kolkata",
            "maxLength": 80
          }
        },
        "required": [
          "title",
          "startsAt",
          "timezone"
        ]
      },
      "UpdateGroupEventDto": {
        "type": "object",
        "properties": {
          "title": {
            "type": "string",
            "minLength": 1,
            "maxLength": 160
          },
          "description": {
            "type": "string",
            "maxLength": 2000
          },
          "location": {
            "type": "string",
            "maxLength": 200
          },
          "startsAt": {
            "type": "string",
            "format": "date-time"
          },
          "endsAt": {
            "type": "string",
            "format": "date-time"
          },
          "timezone": {
            "type": "string",
            "example": "Asia/Kolkata",
            "maxLength": 80
          }
        }
      },
      "UpsertGroupEventRsvpDto": {
        "type": "object",
        "properties": {
          "status": {
            "type": "string",
            "enum": [
              "going",
              "maybe",
              "declined"
            ]
          }
        },
        "required": [
          "status"
        ]
      },
...
    }
  }
}